% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVDplot.R
\name{DVDplot}
\alias{DVDplot}
\title{Plotting DVDtest}
\usage{
DVDplot(tobj, kxlab = NULL, kylab = NULL, kname = NULL,
  thhd = 0.05)
}
\arguments{
\item{tobj}{a return test object of \code{\link{DVDtest}}}

\item{kxlab}{a title for the x axis, \code{.index}}

\item{kylab}{a title for the y axis, \code{.value}}

\item{kname}{a name for \code{k}, e.g. ROI in the references}

\item{thhd}{a number to mark out the zone of the p-value of interest. Defaults to 0.05}
}
\value{
a list of ggplot objects on p value curves and varying distributions
}
\description{
Plot a list of the DVDtest-related figures via \code{ggplot2}
}
\details{
Fig \code{pfig} illustrates a collection of the p value curves among all \code{k}.
Fig \code{kfig} illustrates a collection list of the figures with varying distributions among all 
\code{k}, highlighted the zones of small p values in dark, where \code{ydata1} and 
\code{ydata2} mark as red and blue, repectively. The dashed and dotted lines denote the smooth mean function and +/- 2 
sigma, respectively.
}
\note{
Please contact the maintainer if need more details.
}
\references{
reiss-EMR18.pdf
}
\seealso{
Examples in \code{\link{DVDtest}}
}
\author{
Meng Xu
}
\keyword{ggplot}
