% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVDtest-package.R
\docType{package}
\name{DVDtest-package}
\alias{DVDtest-package}
\title{Difference between Varying Distributions Test (DVDtest)}
\description{
This package contains functions of a test on the difference between varying distributions.
}
\references{
reiss-EMR18.pdf
}
\seealso{
\code{\link{DVDtest}}
}
\author{
Meng Xu, Philip Reiss
}
