% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.params.R
\name{get.params}
\alias{get.params}
\title{Fitting the generalized gamma (Pearson type III) distribution for each roi}
\usage{
get.params(roi, nperm, permarray, eval.index.grid)
}
\arguments{
\item{roi}{a scalar, \code{roi}th data.frame in \code{ydata1&2}}

\item{nperm}{a scalar, number of permutation}

\item{permarray}{an array, permuted-data distances from \code{wass_perm}}

\item{eval.index.grid}{a vector, evaluation grid of \code{.index}}
}
\value{
\item{mu}{a vector of mean}, \item{sigma}{a vector of sigma}, 
\item{nu}{a vector of nu}, \item{aic}{the Akaike information criterion in \code{gamlss}}
}
\description{
Fitting by the permuted-data distances
}
\seealso{
\code{\link{get_params}}
}
\author{
Philip Reiss, Meng Xu
}
