% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pval.R
\name{get.pval}
\alias{get.pval}
\title{Calculate the corrected P values}
\usage{
get.pval(permarray, param.array, realdists, nroi, eval.index.grid, nperm)
}
\arguments{
\item{permarray}{an array, permuted-data distances from \code{wass_perm}}

\item{param.array}{an array, permuted-data distances from \code{get_params}}

\item{realdists}{a matrix or vector, the value from \code{get_realdist}}

\item{nroi}{a scalar, the length of \code{ydata1}}

\item{eval.index.grid}{a vector, evaluation grid of \code{.index}}

\item{nperm}{a scalar, number of permutation}
}
\value{
a vector or matrix of p value
}
\description{
Calculate the corrected P values
}
\seealso{
\code{DVDtest}
}
\author{
Philip Reiss, Meng Xu
}
