% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pval.sd.R
\name{get.pval.sd}
\alias{get.pval.sd}
\title{Calculate the corrected P values with step-down process}
\usage{
get.pval.sd(permarray, param.array, realdists, nroi, grid, nperm)
}
\arguments{
\item{permarray}{an array, permuted-data distances from \code{wass_perm}}

\item{param.array}{an array, permuted-data distances from \code{get_params}}

\item{realdists}{a matrix or vector, the value from \code{get_realdist}}

\item{nroi}{a scalar, the length of \code{ydata1}}

\item{grid}{a vector, evaluation grids of \code{.index}}

\item{nperm}{a scalar, number of permutation}
}
\value{
a vector or matrix of p value
}
\description{
Calculate the corrected P values with step-down process
}
\seealso{
\code{\link{DVDtest}}
}
\author{
Philip Reiss, Meng Xu
}
\keyword{internal}
