% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.realdist.R
\name{get.realdist}
\alias{get.realdist}
\title{Calculating the distances under the null hypothesis for each roi}
\usage{
get.realdist(k, vdFun, ydata1, ydata2, ind.grid, ..., excl, dist.method)
}
\arguments{
\item{k}{a scalar, \code{k}th data.frame in \code{ydata1&2}}

\item{vdFun}{a function, \code{gam} or \code{gamlss}, for fitting the varying distributions}

\item{ydata1}{see \code{DVDtest}}

\item{ydata2}{see \code{DVDtest}}

\item{ind.grid}{see \code{eval.index.grid} in \code{DVDtest}}

\item{\dots}{arguments of \code{vdFun}}

\item{excl}{an argument of \code{predict} for \code{gam}}

\item{dist.method}{see \code{DVDtest}}
}
\value{
a vector of the distances
}
\description{
Calculating the distances under the null hypothesis for each roi
}
\seealso{
\code{DVDtest}
}
\author{
Philip Reiss, Meng Xu
}
