% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Fitting the generalized gamma (Pearson type III) distribution}
\usage{
get_params(nroi, nperm, permarray, eval.index.grid)
}
\arguments{
\item{nroi}{a scalar, the length of \code{ydata1} or \code{ydata2}}

\item{nperm}{a scalar, number of permutation}

\item{permarray}{an array, permuted-data distances from \code{wass_perm}}

\item{eval.index.grid}{a vector, evaluation grid of \code{.index}}
}
\value{
an array, \code{param.array}
}
\description{
Fitting the generalized gamma (Pearson type III) distribution
}
\seealso{
\code{DVDtest}
}
\author{
Meng Xu, Philip Reiss
}
