% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_realdist.R
\name{get_realdist}
\alias{get_realdist}
\title{Calculating the distances under the null hypothesis}
\usage{
get_realdist(vdFun, ydata1, ydata2, grid, ..., excl, mc.cores, dist.method)
}
\arguments{
\item{vdFun}{a function, \code{gam} or \code{gamlss}, for fitting the varying distributions}

\item{ydata1}{see \code{DVDtest}}

\item{ydata2}{see \code{DVDtest}}

\item{grid}{see in \code{DVDtest}}

\item{\dots}{arguments of \code{vdFun}}

\item{excl}{an argument of \code{predict} for \code{gam}}

\item{mc.cores}{a scalar, an argument in \code{mclapply}}

\item{dist.method}{see \code{DVDtest}}
}
\value{
a vector or matrix of the distances and list of parameters of varying 
distributions.
}
\description{
Calculating the distances under the null hypothesis
}
\seealso{
\code{\link{DVDtest}}
}
\author{
Meng Xu, Philip Reiss
}
\keyword{internal}
