% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.perms.R
\name{make.perms}
\alias{make.perms}
\title{Making permutated index}
\usage{
make.perms(dat1, dat2, nperm, .index, adj = permadj)
}
\arguments{
\item{dat1}{an element of \code{ydata1}}

\item{nperm}{a scalar, number of permutation}

\item{.index}{see \code{eval.index.grid} in \code{DVDtest}}

\item{adj}{see \code{permadj} in \code{DVDtest}}

\item{dat1}{an element of \code{ydata2}}
}
\value{
a matrix, permuted indices
}
\description{
Making permutated index
}
\seealso{
\code{DVDtest}
}
\author{
Philip Reiss, Meng Xu
}
