% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiwass.R
\name{multiwass}
\alias{multiwass}
\title{Calculating the distances between two gam/gamlss objects}
\usage{
multiwass(obj1, obj2, newdata1, newdata2, dist.method, ...)
}
\arguments{
\item{obj1}{a gam/gamlss object}

\item{obj2}{another gam/gamlss object}

\item{newdata1}{related evaluation grids}

\item{newdata2}{related evaluation grids}

\item{dist.method}{see \code{DVDtest}}

\item{\dots}{partial arguments in \code{predict}}
}
\value{
a vector, distances
}
\description{
Calculating the distances between two gam/gamlss objects
}
\author{
Philip Reiss, Meng Xu
}
\keyword{internal}
