% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params2qfunc.R
\name{params2qfunc}
\alias{params2qfunc}
\title{Getting the quantile/density function via parameters}
\usage{
params2qfunc(params, family, dist.method)
}
\arguments{
\item{params}{a vector, parameters for certain distributions}

\item{family}{a specific distribution from \code{gam} or \code{gamlss}}

\item{dist.method}{see \code{DVDtest}}
}
\value{
quantile or density functions
}
\description{
Getting the quantile/density function via parameters
}
\author{
Meng Xu, Philip Reiss
}
\keyword{internal}
