% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfuncs2wass2.R
\name{qfuncs2wass2}
\alias{qfuncs2wass2}
\title{Distance functions}
\usage{
qfuncs2wass2(qfunc1, qfunc2, dist.method = dist.method, ...)
}
\arguments{
\item{qfunc1}{quantile or density functions from \code{params2qfunc}}

\item{qfunc2}{quantile or density functions from \code{params2qfunc}}

\item{dist.method}{see \code{\link{DVDtest}}}

\item{\dots}{extra arguments in \code{integrate}}
}
\value{
distance functions
}
\description{
Distance functions
}
\author{
Meng Xu, Philip Reiss
}
\keyword{internal}
