% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{Quite functions}
\usage{
quiet(x)
}
\arguments{
\item{x}{the stuff need to be quiet, messages of which are removed}
}
\value{
quite result
}
\description{
Quite functions
}
\author{
Meng Xu
}
\keyword{internal}
