% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wass.perm.R
\name{wass.perm}
\alias{wass.perm}
\title{Calculating the distances via permuted data for each k}
\usage{
wass.perm(ydat1, ydat2, ..., permat, grid, exclude, dist.method, mgcv.gam)
}
\arguments{
\item{ydat1}{\code{k}th \code{data.frame} of \code{ydata1}}

\item{ydat2}{\code{k}th \code{data.frame} of \code{ydata2}}

\item{\dots}{arguments of the function \code{gamlss} if \code{mgcv.gam = FALSE}}

\item{permat}{a result of \code{make.perm}}

\item{grid}{see \code{grid} in \code{DVDtest}}

\item{exclude}{an argument of \code{predict}}

\item{dist.method}{see \code{DVDtest}}

\item{mgcv.gam}{see \code{DVDtest}}
}
\value{
a matrix of permuted-data distances
}
\description{
Calculating the distances via permuted data for each k
}
\seealso{
\code{wass_perm}
}
\author{
Philip Reiss, Meng Xu
}
\keyword{internal}
