% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVDplot.R
\name{DVDplot}
\alias{DVDplot}
\title{Plotting DVDtest}
\usage{
DVDplot(tobj, kxlab = NULL, kylab = NULL, kname = NULL)
}
\arguments{
\item{tobj}{a return test object of \code{\link{DVDtest}}}

\item{kxlab}{a title for the x axis, \code{.index}}

\item{kylab}{a title for the y axis, \code{.value}}

\item{kname}{a name for \code{k}, e.g. ROI in the references}
}
\value{
a list of ggplot objects on p value curves and varying distributions
}
\description{
Plot a list of the DVDtest-related figures via \code{ggplot2}
}
\details{
This is the Details section
}
\references{
reiss-EMR18.pdf
}
\seealso{
Examples in \code{\link{DVDtest}}
}
\author{
Meng Xu
}
\keyword{ggplot}
