% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVDtest-package.R
\docType{package}
\name{DVDtest-package}
\alias{DVDtest-package}
\title{c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_title(\"#1\")}",
"DVDtest")\Sexpr{tools:::Rd_package_title("DVDtest")}}
\description{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_description(\"#1\")}",
"DVDtest")\Sexpr{tools:::Rd_package_description("DVDtest")}
}
\details{
The DESCRIPTION file:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_DESCRIPTION(\"#1\")}",
"DVDtest")\Sexpr{tools:::Rd_package_DESCRIPTION("DVDtest")}
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_indices(\"#1\")}",
"DVDtest")\Sexpr{tools:::Rd_package_indices("DVDtest")} ~~ An overview of
how to use the package, including the most important functions ~~
}
\references{
reiss-EMR18.pdf
}
\seealso{
\code{\link{DVDtest}}
}
\author{
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_author(\"#1\")}",
"DVDtest")\Sexpr{tools:::Rd_package_author("DVDtest")}

Maintainer:
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_package_maintainer(\"#1\")}",
"DVDtest")\Sexpr{tools:::Rd_package_maintainer("DVDtest")}
}
\keyword{package}
