% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVDtest-package.R
\docType{package}
\name{DVDtest-package}
\alias{DVDtest-package}
\title{Difference between Varying Distributions Test (DVDtest)}
\description{
This package contains a series of internal and external functions of 
Difference between Varying Distributions Test (DVDtest), which tests the pointwise
group differences two varying distributions.
}
\seealso{
\code{\link{DVDtest}}
}
\author{
Meng Xu, Philip Reiss
Maintainer: Meng Xu <mxu@campus.haifa.ac.il>
}
