% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVDtest.R
\name{DVDtest}
\alias{DVDtest}
\title{Difference between Varying Distributions Test (DVDtest)}
\usage{
DVDtest(ydata1, ydata2, nperm, eval.index.grid, dist.method = "wass",
  mgcv.gam = TRUE, ..., exclude = NULL, permadj = FALSE,
  mc.cores = 1)
}
\arguments{
\item{ydata1}{a \code{data.frame} or a \code{list} of \code{data.frame} containing 
at least 3 columns called '\code{.obs}','\code{.index}','\code{.value}' which 
specify which curve the point belongs to (\code{.obs}) at which ('\code{.index}') 
it was observed and the ovserved value (\code{'.value'}). Other variables are 
available for modelling the varying distribution as well.}

\item{ydata2}{same type as \code{ydata1}. If the type of \code{ydata1} and \code{ydata2}
is a \code{list} of \code{data.frame}, the lenghs of two lists must be the same.}

\item{nperm}{a scalar, number of permutation}

\item{eval.index.grid}{a vector, evaluation grid of \code{.index}}

\item{dist.method}{the methods, Wasserstein(\code{'wass'}), L2(\code{L2}), 
L1(\code{'L1'}) and Hellinger(\code{'Hellinger'}) to calculate the distances between 
distributions. Defaults to \code{'wass'}.}

\item{mgcv.gam}{a logical variable, whether to apply \code{mgcv::gam} for eastimating 
distributions, whose parameters are a smooth function of a continuous variable. If 
\code{FALSE}, \code{gamlss::gamlss} is adopted.}

\item{\dots}{If \code{mgcv.gam = TRUE}, \code{\dots} should include \code{formula} and 
\code{family}(=\code{gaulss()}) and other arguments in \code{mgcv::gam}. Otherwise, 
the arguments in gamlss::gamlss are included.}

\item{exclude}{works only in the case of \code{mgcv.gam = TRUE}, to exclude the random effect}

\item{permadj}{a logical variable, whether to adjust the permutated data to cover the entire range,
esp. working in case of sparsity. Defaults to \code{FALSE}}

\item{mc.cores}{the same argument in \code{mclapply} (not work on Windows). Defaults to 1.}
}
\value{
{.index}{a vector, evaluation grid} %% \item{pval}{a vector or matrix of p value}
}
\description{
Testing the difference of two varying distributions.
}
\note{
\itemize{
\item If \code{mgcv.gam} is TRUE and \code{exclue} is NULL (default settings), 
then \code{formula. <- list(.value~s(.index)+s(.obs, bs="re"), ~s(.index))} and 
\code{exclude. <- "s(.obs)"}
}
}
\examples{

 p=6
 mu1<-function(t) 0.2*(p-1)*sin(pi*t)+t+1
 mu2<-function(t) -0.2*(p-1)*sin(pi*t)+t+1
 sig1 <- function(t) t+1
 sig2 <- sig1
 nperson=10
 fun1<-function(t) rnorm(nperson,mu1(t),sig1(t))
 fun2<-function(t) rnorm(nperson,mu2(t),sig2(t))
 tp<-seq(0,1,,10)
 data1<-sapply(tp,fun1)
 data2<-sapply(tp,fun2)
 
 library(reshape2)
 colnames(data1)<-tp
 dg1<-melt(data1)
 colnames(dg1)<-c('.obs','.index','.value')
 dg1$.obs<-as.factor(dg1$.obs)
 
 colnames(data2)<-tp
 rownames(data2)<-1:nperson+2*nperson
 dg2<-melt(data2)
 colnames(dg2)<-c('.obs','.index','.value')
 dg2$.obs<-as.factor(dg2$.obs)
 # library(ggplot2)
 # ggplot()+geom_line(data=dg1,aes(x=.index,y=.value,col=factor(.obs)))+
 #   geom_line(data=dg2,aes(x=.index,y=.value,col=factor(.obs)))

 ngrid=50
 ev.grid <- seq(0, 1, , ngrid)
 nperm. <- 50
 
 simu.test<-DVDtest(dg1,dg2,nperm.,ev.grid)
}
\references{
reiss-EMR18.pdf
}
\author{
Meng Xu, Philip Reiss
}
\keyword{minp}
\keyword{permutation}
