% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.realdist.R
\name{get.realdist}
\alias{get.realdist}
\title{Calculating the distances under the null hypothesis for each k}
\usage{
get.realdist(k, ydata1, ydata2, grid, ..., exclude, dist.method, mgcv.gam)
}
\arguments{
\item{k}{a scalar, \code{k}th data.frame in \code{ydata1&2}}

\item{ydata1}{see \code{DVDtest}}

\item{ydata2}{see \code{DVDtest}}

\item{grid}{see \code{DVDtest}}

\item{\dots}{arguments of \code{vdFun}}

\item{dist.method}{see \code{DVDtest}}

\item{mgcv.gam}{see \code{DVDtest}}

\item{excl}{an argument of \code{predict} for \code{gam}}
}
\value{
a vector of the distances and parameters of varying distributions
}
\description{
Calculating the distances under the null hypothesis for each k
}
\seealso{
\code{\link{DVDtest}}
}
\author{
Philip Reiss, Meng Xu
}
\keyword{internal}
