% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Fitting the Gamma distribution (Pearson type III)}
\usage{
get_params(nroi, nperm, permarray, grid, mc.cores)
}
\arguments{
\item{nroi}{a scalar, the length of \code{ydata1} or \code{ydata2}}

\item{nperm}{a scalar, number of permutation}

\item{permarray}{an array, permuted-data distances from \code{wass_perm}}

\item{grid}{a vector, evaluation grid of \code{.index}}

\item{mc.cores}{a number of the cores using for multi-process}
}
\value{
an array, \code{param.array}
}
\description{
Fitting the Gamma distribution (Pearson type III)
}
\seealso{
\code{DVDtest}
}
\author{
Meng Xu, Philip Reiss
}
\keyword{internal}
