% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wass.perm.R
\name{wass.perm}
\alias{wass.perm}
\title{Calculating the distances via permuted data for each k}
\usage{
wass.perm(np, vdFun, ydat1, ydat2, ..., permat, grid, exclude, dist.method)
}
\arguments{
\item{vdFun}{a function, \code{gam} or \code{gamlss}, for fitting the varying distributions}

\item{\dots}{arguments of \code{vdFun}}

\item{permat}{a result of \code{make.perm}}

\item{grid}{see \code{grid} in \code{DVDtest}}

\item{exclude}{an argument of \code{predict}}

\item{dist.method}{see \code{DVDtest}}

\item{k}{a scalar, \code{k}th \code{data.frame} of \code{ydata1&2}}

\item{dat1}{\code{k}th \code{data.frame} of \code{ydata1}}

\item{dat2}{\code{k}th \code{data.frame} of \code{ydata2}}
}
\value{
a matrix of permuted-data distances
}
\description{
Calculating the distances via permuted data for each k
}
\seealso{
\code{wass_perm}
}
\author{
Philip Reiss, Meng Xu
}
\keyword{internal}
